Move commandline spec
